requirements.txt
flask==2.3.0
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.3.0
google-cloud-bigquery==3.12.0
google-cloud-storage==2.10.0
werkzeug==2.3.0
gunicorn==21.2.0